+++ GEngn/Game/Game.h
@@ -1,4 +1,5 @@
 ﻿#pragma once
+#include "../Engine/Engine.h"
 struct Float2;
 class Player;
 
@@ -6,6 +7,7 @@ class Game
 {
     bool isRunning = false;
     Player* player = nullptr;
+    
     // int PlayerEntityID = -1;
     // int EnemiesToSpawn = 100;
     // std::vector<int> EnemyEntityIDs; 
@@ -14,7 +16,7 @@ class Game
 public:
     void Start();
     void Update();
-    void SpawnBullet(Float2 position, Float2 velocity, int size); 
+    void SpawnBullet(Float2 position, Float2 velocity, float size); 
     void SpawnEnemy(float xPos);
     void NotifyIdChanged(int oldEntityID, int newEntityID);
     void NotifyEntityDestroyed(int eID);
